NDL: do not use 'network drop latency'
Even though our tool is named 'ndl', which comes from network drop latency,
this is not exactly what we collect. In the tool, what we collect is called
"RTD". Let's not use "network drop latency" and use either "RTD" or just
latency. It does not make it much clearer, but at least this is less confusing
and in-line with the documentaion on the web site.

Change-Id: I09c4e38db5df4a7e72180f15eb541e3d000b8747
Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>