drivers: wult: add synthetic events
We are using 'trace_printk()' to send measurement data to user-space. However,
this mechanism is considered to be debug-only and not upstreamable.

This patch adds synthetic event support. Wult driver createes a trace event on
the fily and triggers it, thus sending the data to user-space.

This patch does not remove the 'trace_printk()' mechanism so far, because the
in-kernel synthetic event creation capability was added in kernel version 5.6.
So we'll carry the old mechanism for now. The old stuff is separated using the
'COMPAT_USE_TRACE_PRINTK' macro. We'll remove the old stuff eventually.

Change-Id: I26dfe88f23f246aa0f52fef236c937052964fa54
Signed-off-by: Artem Bityutskiy <dedekind1@gmail.com>