WultRunner: adjust to driver changes
In case of the synthetic events mechansim, trace messages are not prefixed with
"wult: ". And now the driver is changed so that the 'trace_printk()' mechanism
does not add the prefix (this is the only difference between the
'trace_prink()' and synthetic event output).

Now adjust the receiving side - do not expect the prefix.

Change-Id: Ie75029072928e9cc2be91241fc75cb88aea2828e
Signed-off-by: Artem Bityutskiy <dedekind1@gmail.com>