WultRunner: fix error path
When the timer resolution is poor, we are supposed to print a nice message and
exit, but instead, we crash, because we are trying to print an undefined
variable. This patch fixes the problem by initializing the variable earlier.

At the same time, in case of a timer, we should print an additional suggestion
about the CONFIG_HIGH_RES_TIMERS option. But that does not work in case of the
'tdt' device ID. Therefore, check for "timer" word in the description, not in
the device ID.

Change-Id: I87e5d842d9b1424ccd44d89a6685c8cc5e80a2da
Signed-off-by: Artem Bityutskiy <dedekind1@gmail.com>