drivers/wult: fix C-state cycles reporting
Userspace waits for C-state cycles to be reported via tokens like
'CC1Cyc=Value', and this is how it worked with 'trace_printk()'. However, in
commit '49452dc drivers: wult: add synthetic events' we switched to using
synthetic events and mistakenly changed the pair to look like 'CC1=Value'
(notice the missing "Cyc" part). The net result is that userspace drops the
C-state statistics.

This patch fixes the issue.

Change-Id: I944087bed2b4edc2b7777777ed7db755e86b9d6f
Signed-off-by: Artem Bityutskiy <dedekind1@gmail.com>