drivers: wult: rework API for non-nanoseconds drivers
Wult framework provides API to delayed event drivers, and we have 2 of them:
wult_igb and wult_timer. The driver must provide callbacks like
'get_time_before_idle()', and these callbacks can return time in whatever units
the driver prefers.

But we convert those units to nanoseconds before sending the data to userspace.
Since 'wult_igb' driver already operates in nanosecond units, there is nothing
to convert for 'wult_igb'. But the 'wult_timer' driver operates in CPU cycle
units, and we need to convert.

Before this patch, we used a 'unit_is_ns' boolean to distinguish between 'ns'
and 'non-ns' units, and then we assumed the non-ns unit must be CPU cycles.
This is obviously not going to work if there is a 3rd driver with units in, say,
MyClock units.

This patch changes the API: 'unit_is_ns' is removed, and a new callback
'time_to_ns()' is added instead. The 'non-ns' delayed event drivers provide it,
so the framework (tracer.c) should not make any assumptions about the 'non-ns'
driver units.

Change-Id: I04b39ac7e81c920d3ddb11cc0a782d3b8ee70d3f
Signed-off-by: Artem Bityutskiy <dedekind1@gmail.com>