drivers: wult: rename uapi functions
The "uapi" files are from "User-space API", and they contain the
debugfs-related code (should be turned to sysfs at some point later, but
debugfs for now). When a wult device is registered, we create create debugfs
files, and when it is unregistered, we remove them.

This patch renames the debugfs create/remove files:
 * wult_dfs_create() -> wult_uapi_device_register()
 * wult_dfs_remove() -> wult_uapi_device_unregister()

This is more consistent with how we name non-static function
(wult_<subsystem>_) and also more readable in 'main.c', where we call these
functions from 'wult_register()' and 'wult_unregister()'.

Change-Id: I86c7cf0a634176096ba929a2f63bbc1d5fb067df
Signed-off-by: Artem Bityutskiy <dedekind1@gmail.com>