drivers: wult: use less of global 'wi'
We have a global 'struct wult_info wi' object defined in 'main.c', and we
generally try to avoid using the global object, but rather pass it down to
various function calls. This is just a cleaner way than referring to a global
variable.

In 'uapi.c', however, we use the global 'wi' variable, which is inconsistent
with how we do it in other parts of the driver. This patch modifies the
'wult_uapi_device_register()' function to accept a 'wi' pointer so that it
stops using the global 'wi' variable.

Change-Id: Iecb246668ac1b6804e46b10efe21b75568ca4790
Signed-off-by: Artem Bityutskiy <dedekind1@gmail.com>