drivers: wult: simplify enabling/disabling locking
This patch simplifies the locking around enabling and disabling wult
measurements. We enable and disable under the 'enable_lock', and we change the
'wi.enabled' boolean under the same lock. This guarantees that only one process
at a time can enable/disable wult.

But when it comes to reading 'wi.enabled', we do not need the 'enable_lock'
spinlock. We can just read it. Therefore, in the debugfs handler where we read
the 'enabled' status, we also do not need to hold the lock. This patch removes
it and simplifies the code.

Change-Id: Ib0578cb4bef166c3aec00a9f21308bcf3881ae9a
Signed-off-by: Artem Bityutskiy <dedekind1@gmail.com>