drivers: wult: even less globabl 'wi' usage
We do not need to use the globabl 'wi' variable in 'dfs_read_u64_file()', which
handles access to varios read-only wult debugfs file. We have the 'wi' pointer
in 'file->private_data', use it instead.

It is just cleaner and more consistent to not rely on the globabl 'wi'
variable outside of 'main.c'.

Change-Id: I685b30b7f3339ed985b28ab5cbdad6ed64be0251
Signed-off-by: Artem Bityutskiy <dedekind1@gmail.com>