Rename DeployHelpers.py to ToolsCommon.py
There is common code in 'wult' and 'ndl' that is not related to deployment, and
it would be good to have a place for it. We could create a separate module, but
it is more reasonable to just rename the "DeployHelpers.py" file and use it for
any shared code, not only deploy-related shared code.

Change-Id: Ieec45adac9a520f4a3a79d42dbb571362f52fffa
Signed-off-by: Artem Bityutskiy <dedekind1@gmail.com>