wult/ndl: move parts of 'report_command()' to 'ToolsCommon'
Move the code related to opening the raw result to a shared
'report_command_open_raw_results()' function.

This is a preparation to the next patches where we'll add the '--reportids'
option, and there will be more duplicate code, so
'report_command_open_raw_results()' will become larger and sharing the code
will be more justified.

Change-Id: I9dd42ee8b423b3829d49f99f47d62dc634af3de4
Signed-off-by: Artem Bityutskiy <dedekind1@gmail.com>