wult/ndl: add 'report --reportids' option
Very helpful when we want to change report ID for a particular HTML report, and
this happens very often. This commit adds the '--reportids' option for
temporarily overriding the report IDs stored in the 'info.yml' files of the raw
test results.

This patch also updates the man pages with the new option documentation.

Change-Id: I4e5c31a59740f6bc55ae3e49d86a1e433d12d4c7
Signed-off-by: Artem Bityutskiy <dedekind1@gmail.com>