manpages: add vim header
We used vim header everywhere for convenience - when editing with 'vim', it
knows that lines are supposed to be 100 characters long, plus few other things.

Man pages were missing vim headers, and this patch adds them.

At the same time re-format one paragraph that was obviously formatted with 80
characters per line limit, which is inconsistent with the rest of the text in
the man page.

Change-Id: I3a39116f7c8f3a77ece405eddea08a2c14990bc9
Signed-off-by: Artem Bityutskiy <dedekind1@gmail.com>