ProcHelpers: fix killing stale processes
When the process we are killing has a parent process which runs under different
user (e.g., we are killing a 'user' process, but its parent is a 'root'
process), we cannot kill the parent if we run under 'user' as well.

And we crash with 'Operation not permitted'. This is fine so far.

Howerver, often when the child is killed, the parent just terminates itself. So
we need to be patient and re-check.

This patch fixes the 'kill_pids' behavior and make it be a bit more patient.
It already tries to kill the PIDs 2 times, and this patch makes it avoid
erroring out at the 1st attempt - let's do the second attempt and only then
panic if something did not exit.

Change-Id: I196e7bfc115e692a8c37c26b654132e1e5a938c4
Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>