Procs/SSH: do not crash on closed streams
If we run 'wait_for_cmd()' 2 times in a row for a process that closed its
streams, we end up with a crash when reading from a 'None' stream. This patch
fixes the issue.

In addition, make the 'wait_for_cmd()' function in 'Proc' and 'SSH' code be a
bit more similar, since the logic is largely the same.

Change-Id: Ia8bd6b68bf9bddf2a81ddfe707fc9679b15087c0
Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>