Add 'Human' helper module
The entire 'helperlibs' package of modules is coming form another project, but
in a stripped-down format. This is essentially a collection of helpful modules
we share with some other projects. We do not have a separate project for this
shared stuff, so we maintain few copies.

This commit is essentially a sync between several copies. We add the 'Human'
module (even though it is not needed in 'wult'), and this module provides only
a single 'duration()' function, that is used by the _Common module.

Change-Id: I6b166ae94a9d0edc47bb703fd3eed27e09b8e54b
Signed-off-by: Artem Bityutskiy <dedekind1@gmail.com>