RORawResult: rename 'stats' to 'summary', step 1
Today we use term "stats" for 2 different things.

1. The metrics statistics, like median and 99th percentile. These are
   statistics for various metrics in the 'datapoints.csv' file.
2. The SUT statistics that wult does not collect today, but which is somewhat
   supported and if present, referred to in the HTML report. See the
   'stats_path' in '_RawResultBase.py'.

The long term plan is to teach wult collecting statistics in parallel with
measuring C-state latency. For example, wult can collect AC power data and
turbostat output in parallel. Therefore, the small terminology ambiguity is
going to be a problem in the long run.

This patch starts renaming the "stats" term into "summary" when it refers to
CSV file metrics' statistics. In the code we'll try to use short version
'smry', which is hard to read, but it is not as confusing as 'sum', because
'sum' is a common term for addition.

This patch starts the renaming. It renames several functions and variables, as
well as adjusts commentaries and messages.

Change-Id: I39a36042d4ef07a9929b64d863a3ba8261d7de6f
Signed-off-by: Artem Bityutskiy <dedekind1@gmail.com>