Rename: summary table -> intro table
This patch renames internal terminology in '_HTMLReportBase.py'. The very first
table in the report contains things like datapoints count and wult version, and
now it is called the 'intro' table.

The reason for the rename is that we are going to use word 'summary' for
something else soon, and this patch should help keeping the terminology
unambiguous.

Change-Id: Ic93bae78b92060d20fbb21aa398107d52701ec33
Signed-off-by: Artem Bityutskiy <dedekind1@gmail.com>