HTMLReportBase: rename "raw stats" to "stats"
This patch is the first step in cleaning up terminology around "stats". We use
this term for 2 things today:

1. The metrics statistics, like median and 99th percentile.
2. The SUT statistics that wult does not collect today, but which is somewhat
   supported and if present, referred to in the HTML report. See the
   'stats_path' in '_RawResultBase.py'.

The long term plan is to teach wult collecting statistics in parallel with
measuring C-state latency. For example, wult can collect AC power data and
turbostat output in parallel. Therefore, the small terminology ambiguity is
going to be a problem in the long run.

So the plan is to use term "sums" for #1, and stick with "stats" for #2. This
is going to be don in the subsequent patches.

This patch, however, eliminates yet another term - "raw stats". It is basically
the same as "stats", but we use "raw stats" in '_HTMLReportBase'. Let's use
term "stats" as well.

Change-Id: I8a2553897003c5e05ce61a9d60aba6ad58371623
Signed-off-by: Artem Bityutskiy <dedekind1@gmail.com>