ndl: remove '--cstate-stats' support
This option is not very useful because it is implemented completely differently
comparing to C-state data support in wult. In wult, we collect C-state data
inside the kernel and around the 'mwait' instruction. In case of ndl, we use
turbostat to collect C-stae information, and we start it very early, and stop
very late, so the data we get covers huge area around a datapoint, and we found
that this is not really useful, and this slows down ndl a lot.

Remove the support, becaouse no one uses it. Instead, we are planning to add
generic statistics support in the future, which will include turbostat and
more.

Change-Id: Ibbcb79aba996c442b0d9395702f37eb459ef89fc
Signed-off-by: Artem Bityutskiy <dedekind1@gmail.com>