CPUInfo: sync with main project
The CPUInfo module that we use in wult actually comes from a different project
of ours, which is not open-source yet. So we have to carry a copy of CPUInfo in
wult. This patch just synchronizes wult's 'CPUInfo' copy with the main
project. At some point we'll open-source that project and won't have to carry a
copy in wult.

Change-Id: Ib21f0fe1168ff25821c1a020cb604e5d0e9192e6
Signed-off-by: Artem Bityutskiy <dedekind1@gmail.com>