CPUIdle: sync with main project
The CPUIdle module that we use in wult actually comes from a different project
of ours, which is not open-source yet. So we have to carry a copy of CPUIdle in
wult. This patch just synchronizes wult's 'CPUIdle' copy with the main
project. At some point we'll open-source that project and won't have to carry a
copy in wult.

Change-Id: I7293c25ff3036ccce3a438294f71de148b9fbf8a
Signed-off-by: Artem Bityutskiy <dedekind1@gmail.com>