FSHelpers: sync with main project
The 'FSHelpers' module that we use in wult actually comes from a different project
of ours, which is not open-source yet. So we have to carry a copy of
'FSHelpers' in wult. This patch just synchronizes wult's 'FSHelpers' copy with
the main project. At some point we'll open-source that project and won't have
to carry a copy in wult.

Change-Id: I7127e61f202a80a7332fc6df538ce53edc851f7b
Signed-off-by: Artem Bityutskiy <dedekind1@gmail.com>