Move ldist and cpunum parsing to the tools
Move launch distance range ('--ldist' option) string parsing and validation from
the 'WultRunner' module up to the wult tool and make the 'WultRunner' API a bit
simpler, so that it does not assume a string, just a tuple of numbers instead.
Let the tools parse the strings. Similar change was done for ndl and for the
CPU number value ('--cpunum' option)

At the same time simplify the assumptions about the 'ldist' argument:
  * was: a tuple of one or 2 elements or 'None'.
  * now: a tuple of 2 elements or 'None'.

So the change there is that when user provides a single '--ldist' value, we
parse it in a tuple of 2 elements of the same value. This makes processing a
bit simpler.

Change-Id: I4f5900cc584227d0caeed6fa2a3967b161a64e78
Signed-off-by: Artem Bityutskiy <dedekind1@gmail.com>