manpages: do not use backticks
In man pages we escaped most of single quote characters (\'), which resulted in
man showing backtics insteaqd of single quote characters ("`" instead of "'").

Stop doing this. This is more work for no value. We were also inconsistent and
sometimes used sigle quotes for the same stuff where we used backticks.

So this patch removes all the backtics, let's use single quotes instead.

Change-Id: Ic0b023a33008f22d6bc4af257c6b45c6c2c17fd0
Signed-off-by: Artem Bityutskiy <dedekind1@gmail.com>