RORawResult: improve error messages
In some cases we deal with broke test report directories that either have not
info/datapoint file, or those files are empty. This may happen when wult test
was interrupted.

These cases are already handled and wult exits with an error, but the error
message is not super clear. Let's improve this by checking for these error
cases explicitly.

This is not a bug fix, just a non-essential improvement.

Change-Id: I39ec9d136f8dd3f97d68135df25b32e21eb81f97
Signed-off-by: Artem Bityutskiy <dedekind1@gmail.com>