Stop using OrderedDict
In many places we like to preserve order, and this is why we use 'OrderedDict'.
However, starting from python 3.6 normal dictionaries also preserve the order
(guaranteed starting from python 3.7).

Therefore, we do not really need to use 'OrderedDict' any longer. It would be
OK to still use it for increased readability, but the problem is that it is
inconvenient when one wants to print it or dump some data structures including
it - many just cannot nicely print 'OrderedDict' type in a pretty way. Using
plain dictionary is an advantage in this case.

Since wult requires Python 3.6+ anyway, we can safely dump 'OrderedDict' now.

Change-Id: Ib0daf159f2e81e4e79abc97e2b14be78440fc879
Signed-off-by: Artem Bityutskiy <dedekind1@gmail.com>