CPUIdle: fix crash
We releasec wult version 1.8.11, but it is completely broken because we forgot
to add the MSR module. We had it on our system, so we never noticed the bug,
but when wult is installed on a new system, nothing works because the module
is not there.

Add the module to fix the issue.

Change-Id: I37861c7c9380444184fece7b03aae58452c25c9f
Signed-off-by: Artem Bityutskiy <dedekind1@gmail.com>