ProcHelpers: fix waiting for child processes
We use 'kill -0' to check if child processes exited. And then we check for
'kill' exitcode - if it is non-zero, we assume all processes exited.

But this is wrong, 'man kill' tell sus that we should check for exit code '1'
instead, which means that signal was not delivered to any process, which is
what we need, because that means all of them exited. There is exit code '64',
which means that some signals were not delivered, and this is not good enough
for us.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@intel.com>