MSR: streamline register comments
We add more and more MSR definition, and it becomes harder to visually
distinguish between definitions for different MSRs. Add a header comment before
each definition. Not that it brings new information, but it helps to visually
distinguish between different MSR blocks.

Remove the MSR_TURBO_GROUP_CORECNT definition as we do not use it.

Change-Id: I6d126e590dcae56effb2b9e0cad965b14c08264f
Signed-off-by: Artem Bityutskiy <dedekind1@gmail.com>