wult/ndl: deprecate --continue and add --start-over
By default wult and ndl truncate the CSV file and throws away potentially
valuable datapoints. This is a poor choice for default behavior, we do not want
to delete user data if not explicitely asked.

In order to fix this, this patch deprecates the '--continue' option. It is
still there, but it is hidden and has no effect. Instead, the "continue"
behavior is now the default. The '--start-over' option should be used in order
to override the pre-existing datapoints.

Change-Id: Iac9522691b988b6a3fbfc3bf744ca203176c7e64
Signed-off-by: Artem Bityutskiy <dedekind1@gmail.com>