ArgParse: add a workaround for man pages
For some reason 'argparse-manpage' fails to handle subcommand descriptions and
does not remove newlines from them. For example, this is how the "filter"
subcommand description looks like:

Filter datapoints out of a test result by removing CSV rows and columns according to
	       specified criteria. The criteria is specified using the row and column filter and
	       selector options ('--rsel', '--cfilt', etc). The options may be specified multiple
	       times.

With this patch, it looks looks like this.

Filter  datapoints  out of a test result by removing CSV rows and columns
according to specified criteria. The criteria is specified using the row and
column filter and selector options ('--rsel', '--cfilt', etc). The options may
be specified multiple times.

Change-Id: I57aeed144df1c2466aa9c8177f4a8bcc5c77b94a
Signed-off-by: Artem Bityutskiy <dedekind1@gmail.com>