driver: capture interrupt time-stamp
This is the first step in implementing interrupt latency tracing. Make the igb
and timer drivers capture time-stamp in the interrupt handler and pass it to
'wult_interrupt()'. We do not do anything with it for now, but the plan is to
send it to user-space along with other tracing data.

Change-Id: Ie8abc83ad3a42141f218656bf21358a3fb06c2d7
Signed-off-by: Artem Bityutskiy <dedekind1@gmail.com>