wult: fix running by PCI ID
Fix a falure with the following symptoms:

$ wult start 0000:18:00.0
wult: error: network interface '0000:18:00.0' was not found.
Here is the list of network iterfaces available: eno1, eno2.

The network interface corresponding to the NIC with PCI ID '0000:18:00.0'
indeed does not exist, because it was unbinded from the standard 'igb' driver.
Them may happend, for example, because previous wult run was killed and wult
had no chance to re-bind the '0000:18:00.0' NIC.

But we should anyway accept the valid, existin device '0000:18:00.0'. We should
not fail if there is no network interface for it, we do not really need the
network interface.

This patch addresses the issue.

Change-Id: I873e391c2d267982a99b14c26c5ea5fc5b332436
Signed-off-by: Artem Bityutskiy <dedekind1@gmail.com>