drivers: tracer: fix warnings on old kernels
During tracepoint register and unregister times the 'cpu_idle_hook()' is called
with an unexpected pattern. Namely, during registering, it is called with
'PWR_EVENT_EXIT' several times in a row (like we keep exiting a C-state several
times without entering it). During registering, it is called with a
non-'PWR_EVENT_EXIT' value several times in a row, like we keep entering a C-state
without exiting. This is some sort of tracing subsystem artifact, or bug. We do
not observe it in new kernels.

As a result, with old kenels, there are warnings coming from
'WARN_ON(before_idle_called)' in 'tracer.c'. There are no warning starting from
kernel version 5.3.

This commit makes that warning to be conditional on the kernel version. This is
ugly, but temporary. We'll drop old kernels support at some point.

Change-Id: I04ea3e7a25e56b8110aa46656138b3c90cb6c451
Signed-off-by: Artem Bityutskiy <dedekind1@gmail.com>