wultlibs: set attributes to 'None' in close
This patch is just hardening. I saw really strange crashes happening in
'close()' in various places, often times the crashes are related to the fact
that our 'close()' can be called two times - directly by user and then from
'__del__()'. In order to avoid double de-initialization, we should set all the
attributes that we deinitialized to 'None'.

Change-Id: I9de7fc3367fd00bb4740825b67cd31d345b3030e
Signed-off-by: Artem Bityutskiy <dedekind1@gmail.com>