wult deploy: prepare to support multiple helpers
We are going to add the statistics collector helper to both wult and ndl.
However, today we assume that every tool (wult, ndl) has its own set of
helpers, which are not shared. So we'd have to add the statsitics collection
helper to both 'wult' and 'ndl', so have two copies.

This patch is a preparation work that is trying to change this model.

Instead of assuming that ndl helpers are in 'helpers/ndl', while wult helpers
are in 'helpers/wult', we will assume that helpers are in
'helpers/<helper_name>'. This way we'll be able to use the same helper for both
tools.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>