drivers: wult: further optimize 'after_idle()'
We measure the overhead of 'after_idle()' by taking TSC at the beginning and at
the end. But in case of 'wult_timer' driver, the driver also just returns TSC
in from 'get_time_after_idle()'. This is not optimal.

In order to optimize this, we just remove 'get_time_after_idle()' from the
'wult_timer' method, and introduce the assumption that if the method is not
defined, then time after idle should be taked by just reading TSC. This way we
end up reading TSC only once.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>