drivers: wult_igb: fix adjustment in IRQ handler
When we take the IRQ handler time, we read the 'I210_SYSTIMR' register, which
takes time. We are trying to adjust for the read overhead, but do it
incorrectly. We should assume that we reach the I210 HW roughly in the middle
of the read interval. Therefore, use half of the read time in the adjustment,
not the entire read time.

In additional, make sure we flush posted PCI writes before reading the
'I210_SYSTIMR' register.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>