Devices: simplify TDT class
Remove the base class for _TSCDeadlineTimer - we do not really need it because
there is only one child class. We are going to introduce an hrtimer class, but
it is not going to share anything with _TSCDeadlineTimer. Therefore, there is
no reason for the base class - it only increases complexity.

Also remove the abstract class stuff. We do not need it, all it did for us was
silencing pylint. Let's siclence it a simpler and more direct way - using the
"pylint: disable=" statement.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>