ToolsCommon: simlify deploy #1: adjust help messages
This patch is part of the series that simplifies 'wult deploy' by removing the
"build host" and "install host" concepts. They are not used, they have been
broken for quite some time, and they make code more complex without much value
added.

In this patch we change 'wult deploy' command line option help messages. We
make them clearly say that we always build drivers and helpers on the SUT.

Use standard 'add_ssh_options()' function fo adding SSH options. In the past,
we did not use it because we had to distinguish between the build and install
hosts. But since we are going to remove this distinction, we can just use the
common function.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>