ToolsCommon: restructure deploying code
We are going to add another class of deployables - python helpers. They need to
be treated differently to helpers. Let's prepare the deployement code for this.

Restructure the code so that different classes of deployables have their own
deploy function: _deploy_helpers() and _deploy_drivers(). This is also easier
to understand and maintain.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>