RemoteHelpers: new helper module
There is a need for a new module where we can put all the unified functions
that work both on local and remote systems. Create new 'RemoteHelpers' module
for this. Add the following functions there:

1. get_free_port() - moved from ProcHelpers, since it is not about processes.
2. time_time() - emulates python 'time.time()' for a remote system.

ToolsCommon is also modified to sue 'RemoteHlpers.time_time()'. In addition to
thise, there are few minor cleanups in ToolsCommon.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>