ToolsCommon: re-implement python helpers deployment
Re-impliment 'stats-collect' deployment. It did not work correctly before.

The main change is that now we always build the stand-alone script version on
the controller. In the past we were doing this on the SUT, which did not always
work because the SUT may not have all the required python dependencies.

The other big change is that we moved standalone script creation from the shell
script to ToolsCommon.py.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>