wult deploy: remove --kmod-path
Remove the 'wult deploy --kmod-path' option. This option specifies path to
deploy drivers to. But this option is not really useful because there is only
one location where we can install the drivers. This location depends on the OS,
but for major distribution it is /lib/modules/<kernel version>.

This option can theoretically be useful when installing on an unsupported OS,
but it also adds complexity, and we never test it. In case of a new OS, it is
better to just learn the right path to install automatically.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>