wult deploy: remove --helpers-path
Yet another simplification - remove the unused '--helpers-path' option. And
again for the same reason - it does not really work with 'is_deploy_needed()'
function - if helpers were installed to a non-standard place, the function will
not know it and won't do the check correctly.

We do not really need this option. We have the 'WULT_HELPERSPATH' environment
variable mechanism for this.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>