Procs/SSH: clean-up 'shell=True' usage
In most cases where we use run a command with 'shell=True' we do it for the
sake of 'SSH', which has 'shell=False' by default, and requires 'shell=True' in
order to figure out the PID, which we then use to monitor/kill the process.

But now 'SSH' has 'shell=True' by default, and it is better to just let it use
the default value.

Indeed, it is not a good idea for users of 'SSH' to rely on 'shell=True' be
"provide me the PID". It is better if 'SSH' just provides the PID by default.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>