SSH: implement interactive shell support - part 1
Start adding interactive shell support. The idea is to run commands in an
interactive shell instead of creating new SSH session for every command - this
should be faster.

This patch just adds some of the required infrastructure, without actually
implementing interactive shell support. It should be easy to review this way.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>