Procs/SSH: clean up private attributes
We add private attributes to the process objects, and in order to avoid possible
conflics, we name them as "_something_". There are many of the private attributes,
and we are going to add more. This is becoming a bit ugly.

This patch moves most of those private attributes to a separate object. I think
this makes code a bit cleaner and more readable.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>