Procs/SSH: fix 'by_line'
When 'by_line' is used, we should not provide partial lines unless the command
exited.

In order to guarantee this, we need to save the partial lines that we read from
the queue, but did not send to the user, in the private data. And return
partial line only when the project exits.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>