SSH: fix occasional crash
Sometimes there is some 'stderr' output in the queue after the "end of process"
marker has been seen. And we may occasionally return the exit code to the user,
even though we did not consume some of the 'stderr' output. This triggers our
assert:

assert pd.queue.empty()
AssertionError

Fix this by exhausting the queue before returning to the user.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>