SSH: save and print the real command in debug messages
The real command that we execture over SSH is a little bit different to the
user-specified command, because we add prefixes/suffixes for printing PID and
exit status.

Save the real command in private data along with the user-specified command,
and print it in debug messages. This is very helpful when debugging issues.

In addition, fix few mistakes in debug messages.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>