Procs/SSH: rework and fix 'wait_for_cmd()'
Was: 'wait_for_cmd()' returns all the output of the program if it finishes,
regarless of the 'lines=(a, b)' limit.

Now: 'wait_for_cmd()' returns only the necessary output even if the program
finished. It will delay returning the exitcode in this case.

Why: this is necessary to fix 'read_pid()' in order to make sure it always
reads only the first line, and even if the program finishes, it never gets the
exitcode. The user will get the exit code when invoking 'wait_for_cmd()'.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>