Procs/SSH: minor code restructure
Restructure the loop in '_do_wait_for_cmd()'. At this point it makes no real
change, but will be needed later, so this is a preparation.

In the upcoming changes we'll make 'wait_for_cmd()' return partial output even
when the command has finished and we got the exit code already. So it is
expected that the user runs 'wait_for_cmd()' again and again, until it gets an
exitcode. With this change, if the exit code is already available and there is
no penidng data in the queu, we wont' even enter the inner loop.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>