Procs/SSH: fix the problem of being stuck
We did not check our output cache at the very beginning, and only checked it
after waiting for new data in the queue. But if new data never arrives, because
the process does not have any new data, and it also does not exit, we never
return the buffered data to the user.

This patch fixes the problem.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>