wult: report: drop 'WakeLatency' for states like 'POLL'
Idle states like 'POLL' have interrupts enabled, so the CPU jumps right to the
interrupt handler, instead of continuing executing cpuidle subsystem
instructions. In this case 'WakeLatency' becomes sort of meaningless, and the
driver does not even measure it, it just sends the same value as for
'IntrLatency'.

This patch ipmroves wult report: if 'WakeLatency' is same as 'IntrLatency', it
excludes 'WakeLatency' from the report.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>