wult/ndl: refuse to continue collecting data
For a long time we supported a feature when we continue collecting data to an
existing test result. But this does not work anymore because of statistics. We
cannot continue collecting statistics. Well, we technically could, but that
would be too difficult to do for very little gain.

Therefore, remove the --start-over and --continue options and always refuse a
directory that already contains wult test results.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>