wult report: fix crash when a test result has only one datapoint
When we generate a diff of 2 results, and one of them has 1 datapoint, while
the other has multiple, we crash, because we try to access the summary table
dictionary for the 'std' function, but it is absent for the result with only
one datapoint.

Traceback (most recent call last):
  File "/opt/powerlab/bin/wult", line 504, in <module>
    sys.exit(main())
  File "/opt/powerlab/bin/wult", line 494, in main
    args.func(args)
  File "/opt/powerlab/bin/wult", line 476, in report_command
    rep.generate()
  File "wultlibs/reportlibs/_HTMLReportBase.py", line 715, in generate
    self._generate_report()
  File "wultlibs/reportlibs/_HTMLReportBase.py", line 238, in _generate_report
    smrys_tbl = self._prepare_smrys_tables(pinfos)
  File "wultlibs/reportlibs/_HTMLReportBase.py", line 149, in _prepare_smrys_tables
    val = res.smrys[colname][funcname]
KeyError: 'std'

This patch fixes the issue.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>