wult: driver: save cycle counters
In 'after_idle()' and the interrupt handler, we use TSC cycle counters for
measuring the overhead. The cycles are stored in local variables. Move them to
'wult_tracer_info' structure.

This is part of a rework where we'll try to just collect everything in
'after_idle()' and the interrupt handler, and then decide on whether the
collected data are valid later, in 'wult_tracer_send_data()'. In other words,
we'll try to do all the calcuations and decisions in 'wult_tracer_send_data()'.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>