wult: driver: save interrupts status
Save the interrupt status (enabled/disabled) in 'wult_tracer_info'. We'll need
it a bit later.

This is part of a rework where we'll try to just collect everything in
'after_idle()' and the interrupt handler, and then decide on whether the
collected data are valid later, in 'wult_tracer_send_data()'. In other words,
we'll try to do all the calcuations and decisions in 'wult_tracer_send_data()'.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>