wult: driver: improve interrupts tracking
We track the intrrupts we arm using the 'intr_happened' boolean variable.
Before this patch, we reset this variable to 'false' every time we entered
idle, which is not very logical.

This patch changes the logic so that we reset the variable to 'false' when we
arm the interrupt.

Also this patch adds a 'WARN_ON()' in case we see an interrupt that we did not
arm.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>