WultRunner: re-use dmesg snapshot
The 'dev' object includes the 'dmesg' snapshot. We already spent time calling
'dmesg' to capture this snapshot. We do not need to do it again in
'WultRunner', we can just re-use 'dev.dmesg_before' instead.

Before this patch, we invoke 'dmesg' 6 times (with --stats none).
After this patch, we invoke it 5 times.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>