KernelModules: reuse 'dmesg' snapshot
Add the 'dmesg' argument to 'KernelModule', similar to the one we have in
'Device.py', but also allow to pass the 'Dmesg' object. This allows to invoke
the 'dmesg' tool less often.

Before this patch, we invoke 'dmesg' 5 times (with --stats none).
After this patch, we invoke it only 1 time.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>