wult: driver: separate data colletion and analysis
Before this patch, we tried to analyze the situation at the same time as we
collected data. E.g., in 'after_idle()', we tried to figure out whether the
interrupt happened or not, and decide whether we want to continue collecting
data or not.

After this patch: just always collect data in 'after_idle()' and the interrupt
handler. Figure out on whether the data are valid or not later, in
'wult_tracer_send_data()'. This simplifies the logic.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>