WultRunner: yield dictionaries from '_get_datapoints()'
Let's yield the raw datapoint dictionary, instead of "fields" and "values"
lists. We anyway later translate the result to a dictionary, so let's do it
earlier, which helps simplifying the code a bit.

Note, we'll later keep raw datapoint and CSV datapoint in different
dictionaries, instead of modifying the rawdp into dp in-place. This is why
there is this strange line: dp = rawdp. We'll do more modifications around this
later.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>