WultRunner: remove the need for 'lscpu' info
The only reason we need the 'lscpu_info' data is to detect if we are dealing
with an Intel CPU. And the only reason we need this, is to figure out if we
have C-state data in raw datapoints. But we can check this just by looking for
a "CCxCyc" field in the raw datapoint. If one if present, we do have C-states.

So simplify the 'WultRunner' class and remove the 'lscpu_info' argument.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>