drivers: wult: remove unneeded assignment
Before running 'before_idle()', we set 'ti->got_dp_ai = false', but this is not
necessary. There is a comment on why we do this, but it is not very helpful.

The reason we do this is that we try to keep only one of the 'got_dp_*' flags
true, but never both. But there is no reason for this, so stop making this
assumption and remove the corresponding 'WARN_ON()'. This makes the code easier
to read.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>