wult: driver: do not sent absent C-states to userspace
Before this patch, wult driver was sending all C-states data to user-space,
even if the C-state was found to not exist on the platform (the counter is
alwasy 0 for an absent C-state). The user-space then would have to filter them
out.

With this patch, if the C-state is absent on this platform, the driver won't
send its data to userspace. This is more optimal.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>